+++ ARMeilleure/Translation/PTC/Ptc.cs
@@ -523,7 +523,7 @@ namespace ARMeilleure.Translation.PTC
             ConcurrentDictionary<ulong, TranslatedFunction> funcs,
             IMemoryManager memory,
             EntryTable<uint> countTable,
-            AddressTable<uint> funcTable,
+            AddressTable<ulong> funcTable,
             TranslatorStubs stubs)
         {
             if (AreCarriersEmpty())
@@ -660,7 +660,7 @@ namespace ARMeilleure.Translation.PTC
             RelocEntry[] relocEntries,
             IntPtr pageTablePointer,
             EntryTable<uint> countTable,
-            AddressTable<uint> funcTable,
+            AddressTable<ulong> funcTable,
             TranslatorStubs stubs,
             out Counter<uint> callCounter)
         {
@@ -790,7 +790,7 @@ namespace ARMeilleure.Translation.PTC
             ConcurrentDictionary<ulong, TranslatedFunction> funcs,
             IMemoryManager memory,
             EntryTable<uint> countTable,
-            AddressTable<uint> funcTable,
+            AddressTable<ulong> funcTable,
             TranslatorStubs stubs)
         {
             var profiledFuncsToTranslate = PtcProfiler.GetProfiledFuncsToTranslate(funcs);
